requirements.txt
# Python dependencies for wildfire detection system
# Requires Python 3.9 or higher

# ===== Core Web Framework =====
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# ===== Database =====
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
geoalchemy2>=0.14.0
alembic>=1.12.0

# ===== Geospatial =====
h3>=3.7.0
# shapely - Optional, only needed if doing complex geometry operations

# ===== Data Processing =====
pandas>=2.1.0
numpy>=1.26.0

# ===== Machine Learning =====
scikit-learn>=1.3.0
joblib>=1.3.0

# ===== HTTP & API =====
requests>=2.31.0

# ===== Utilities =====
python-dotenv>=1.0.0

# ===== Optional: Scheduling =====
# apache-airflow>=2.7.0  # Uncomment if using Airflow for scheduling
# celery>=5.3.0  # Uncomment if using Celery for task queue


